[Amit] Moved managers and strategies packages out of model package and directly under android package
