+++ app/src/main/java/com/mymeetings/android/ViewAlertType.kt
@@ -1,4 +1,4 @@
-package com.mymeetings.android.model.strategies
+package com.mymeetings.android
 
 enum class ViewAlertType {
     RUNNING,
