+++ app/src/main/java/com/mymeetings/android/injections/InjectionsProvider.kt
@@ -5,10 +5,10 @@ import com.mymeetings.android.BuildConfig
 import com.mymeetings.android.db.CalendarEventsReminderDatabase
 import com.mymeetings.android.db.repositories.RoomCalendarEventsDataRepository
 import com.mymeetings.android.debug.ConsoleLog
-import com.mymeetings.android.model.managers.CalendarEventAlertManager
-import com.mymeetings.android.model.managers.CalendarEventsSyncManager
-import com.mymeetings.android.model.strategies.GoogleCalendarFetchStrategy
-import com.mymeetings.android.model.strategies.LocalCalendarFetchStrategy
+import com.mymeetings.android.managers.CalendarEventAlertManager
+import com.mymeetings.android.managers.CalendarEventsSyncManager
+import com.mymeetings.android.strategies.GoogleCalendarFetchStrategy
+import com.mymeetings.android.strategies.LocalCalendarFetchStrategy
 import com.mymeetings.android.utils.ClockUtils
 import com.mymeetings.android.view.viewModels.CalendarEventsViewModel
 import com.mymeetings.android.view.widgets.CalendarEventWidgetRemoteViewFactory
@@ -33,7 +33,11 @@ class InjectionsProvider(private val context: Context) {
 
         single { GoogleCalendarFetchStrategy() }
 
-        single { LocalCalendarFetchStrategy(context) }
+        single {
+            LocalCalendarFetchStrategy(
+                context
+            )
+        }
 
         single {
             //Enabling log only for debug builds.
